feat: enhance PreSwapInfoGroup and SwapServiceFeeOverview components with detailed service fee display and protocol fee comparison
