+++ packages/kit/src/views/Swap/components/SwapServiceFeeOverview.tsx
@@ -1,26 +1,9 @@
-import { useCallback, useMemo } from 'react';
-
 import { useIntl } from 'react-intl';
 
-import {
-  Icon,
-  Image,
-  Popover,
-  SizableText,
-  Stack,
-  XStack,
-} from '@onekeyhq/components';
-import type { IImageProps } from '@onekeyhq/components';
+import { Icon, Popover, SizableText, Stack } from '@onekeyhq/components';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
-import { otherWalletFeeData } from '@onekeyhq/shared/types/swap/SwapProvider.constants';
 
-interface IProtocolFeeInfo {
-  name: string;
-  fee: number;
-  color: string;
-  icon: IImageProps['source'];
-  maxFee: number;
-}
+import { ProtocolFeeComparisonList } from './ProtocolFeeComparisonList';
 
 export function SwapServiceFeeOverview({
   onekeyFee,
@@ -28,48 +11,7 @@ export function SwapServiceFeeOverview({
   onekeyFee: number | undefined;
 }) {
   const intl = useIntl();
-  const renderProtocolFeeListItem = useCallback(
-    (item: IProtocolFeeInfo) => (
-      <XStack gap="$3" alignItems="center">
-        <Stack w={20} h={20}>
-          <Image source={item.icon} w={16} h={16} />
-        </Stack>
-        <Stack flex={1}>
-          <Stack
-            backgroundColor={item.color}
-            borderRadius="$full"
-            width={`${item.maxFee > 0 ? (item.fee / item.maxFee) * 100 : 0}%`}
-            height="$1"
-          />
-        </Stack>
-        <SizableText
-          size="$bodySm"
-          color={item.name === 'oneKey' ? '$textSuccess' : '$text'}
-          textAlign="right"
-        >
-          {item.fee}%
-        </SizableText>
-      </XStack>
-    ),
-    [],
-  );
-
   const serviceFee = onekeyFee ?? 0.3;
-
-  const protocolFeeInfoList: IProtocolFeeInfo[] = useMemo(
-    () => [
-      ...otherWalletFeeData,
-      {
-        maxFee: 0.875,
-        name: 'oneKey',
-        fee: serviceFee,
-        // color: '#202020',
-        color: '#44D62C',
-        icon: require('@onekeyhq/kit/assets/logo.png'),
-      },
-    ],
-    [serviceFee],
-  );
   return (
     <Popover
       title={intl.formatMessage({
@@ -103,9 +45,7 @@ export function SwapServiceFeeOverview({
               )}
             </SizableText>
           </Stack>
-          <Stack gap="$2">
-            {protocolFeeInfoList.map((item) => renderProtocolFeeListItem(item))}
-          </Stack>
+          <ProtocolFeeComparisonList serviceFee={serviceFee} />
         </Stack>
       }
     />
